# app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "iSECTECH Notification & Trust Score API",
    "description": "Production-grade API for notification management and trust scoring systems supporting mobile PWA and enterprise security operations",
    "version": "1.0.0",
    "contact": {
      "name": "iSECTECH API Team",
      "email": "api@isectech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://isectech.com/license"
    }
  },
  "servers": [
    {
      "url": "https://api.isectech.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://api-staging.isectech.com/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Retrieve notifications with filtering and pagination",
        "description": "Get notifications with comprehensive filtering, sorting, and pagination support. Supports up to 1000 requests per minute.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by notification status",
            "schema": {
              "type": "string",
              "enum": ["sent", "delivered", "read", "failed"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by notification type",
            "schema": {
              "type": "string",
              "enum": ["security", "alert", "info", "warning", "error"]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Filter by priority level",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "critical"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results to return",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "notifications": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Notification"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    },
                    "metadata": {
                      "$ref": "#/components/schemas/ResponseMetadata"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Create new notification",
        "description": "Create a new notification for delivery to specified recipients. Supports up to 100 creates per minute.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Notification"
                    },
                    "metadata": {
                      "$ref": "#/components/schemas/ResponseMetadata"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/notifications/templates": {
      "get": {
        "tags": ["Notification Templates"],
        "summary": "Get notification templates",
        "description": "Retrieve notification templates with filtering",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["security", "alert", "info", "warning", "error", "system"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["push", "email", "sms", "webhook"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Templates retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "templates": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/NotificationTemplate"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Notification Templates"],
        "summary": "Create notification template",
        "description": "Create a new notification template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created successfully"
          }
        }
      }
    },
...
    "/notifications/analytics": {
      "get": {
        "tags": ["Notification Analytics"],
        "summary": "Get notification analytics",
        "description": "Retrieve comprehensive analytics for notifications",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Type of analytics to retrieve",
            "schema": {
              "type": "string",
              "enum": ["overview", "detailed", "performance", "engagement", "channels", "trends", "insights"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationAnalytics"
                    },
                    "type": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "notif_123"
          },
          "title": {
            "type": "string",
            "example": "Security Alert"
          },
          "message": {
            "type": "string",
            "example": "Suspicious login attempt detected"
          },
          "type": {
            "type": "string",
            "enum": ["security", "alert", "info", "warning", "error"],
            "example": "security"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"],
            "example": "high"
          },
          "status": {
            "type": "string",
            "enum": ["sent", "delivered", "read", "failed"],
            "example": "delivered"
          },
          "recipients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "channel": {
                  "type": "string",
                  "enum": ["push", "email", "sms", "webhook"]
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateNotificationRequest": {
        "type": "object",
        "required": ["title", "message", "type", "priority", "recipients"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "type": {
            "type": "string",
            "enum": ["security", "alert", "info", "warning", "error"]
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"]
          },
          "recipients": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["userId", "channel"],
              "properties": {
                "userId": {
                  "type": "string"
                },
                "channel": {
                  "type": "string",
                  "enum": ["push", "email", "sms", "webhook"]
                },
                "metadata": {
                  "type": "object"
                }
              }
            }
          },
          "scheduledFor": {
            "type": "string",
            "format": "date-time"
          },
          "templateId": {
            "type": "string"
          },
          "templateData": {
            "type": "object"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                },
                "action": {
                  "type": "string"
                },
                "url": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          }
        }
      },
...
      "NotificationTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "category": {
            "type": "string",
            "enum": ["security", "alert", "info", "warning", "error", "system"]
          },
          "type": {
            "type": "string",
            "enum": ["push", "email", "sms", "webhook"]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          },
          "variables": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": ["string", "number", "date", "boolean", "object", "array"]
                },
                "required": {
                  "type": "boolean"
                },
                "defaultValue": {},
                "description": {
                  "type": "string"
                }
              }
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NotificationAnalytics": {
        "type": "object",
        "properties": {
          "overview": {
            "type": "object",
            "properties": {
              "totalNotifications": {
                "type": "integer"
              },
              "deliveryRate": {
                "type": "number"
              },
              "readRate": {
                "type": "number"
              },
              "clickRate": {
                "type": "number"
              }
            }
          },
          "breakdowns": {
            "type": "object",
            "properties": {
              "byType": {
                "type": "object"
              },
              "byPriority": {
                "type": "object"
              },
              "byChannel": {
                "type": "object"
              }
            }
          },
          "timeSeries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                },
                "sent": {
                  "type": "integer"
                },
                "delivered": {
                  "type": "integer"
                },
                "read": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
...
      "ResponseMetadata": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "req_123456789"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "processingTime": {
            "type": "integer",
            "description": "Processing time in milliseconds",
            "example": 150
          },
          "cached": {
            "type": "boolean",
            "description": "Whether the response was served from cache"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid query parameters"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                },
                "requestId": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Insufficient permissions"
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too Many Requests - Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Rate limit exceeded"
                },
                "retryAfter": {
                  "type": "integer",
                  "example": 60
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal server error"
                },
                "requestId": {
                  "type": "string"
                },
                "details": {
                  "type": "string",
                  "description": "Error details (only in development mode)"
                }
              }
            }
          }
        }
      }
    }
  }
}